Fix wrong data type

+++ audio-client/src/main/java/info/u_team/voice_chat/audio_client/micro/MicroRecorder.java
@@ -45,7 +45,7 @@ public class MicroRecorder implements NoExceptionCloseable {
 		microData.flush();
 	}
 	
-	public void setVolume(double volume) {
+	public void setVolume(int volume) {
 		microData.setVolume(volume);
 	}
